Move shared test setup to unit/int/e2e setups
